move SearchForm functions to separate helpers file
